pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "allmeans"
version = "2.0.0"
authors = [
    { name="Kai Maurin-Jones" },
]
description = "Automatic topic discovery with zero LLMs, minimal input"
readme = "README.md"
requires-python = ">=3.10"
license = { file="LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "scikit-learn>=1.5.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "polars>=1.5.0",
    "gensim>=4.3.0",
    "hdbscan>=0.8.38",
    "umap-learn>=0.5.6",
    "jellyfish>=1.0.0",
    "nltk>=3.8.1",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "matplotlib>=3.9.0",
    "pyyaml>=6.0.0",
    "beautifulsoup4>=4.14.2",
]

[project.optional-dependencies]
sentiment = [
    "nltk>=3.9.0",
    "vaderSentiment>=3.3.0",
]
embeddings = [
    "sentence-transformers>=2.2.2",
    "gensim>=4.3.0",
]
viz = [
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
]
all = [
    "allmeans[sentiment,embeddings,viz]",
]
dev = [
    "allmeans[sentiment,embeddings,viz]",
    "build>=1.0.0",
    "twine>=6.0.0",
    "hatchling>=1.18.0",
    "ruff>=0.8.0",
    "mypy>=1.11.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.5.0",
    "pre-commit>=4.0.0",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
]

[project.scripts]
allmeans = "AllMeans.cli:app"

[project.urls]
Homepage = "https://github.com/kmaurinjones/AllMeans"
Documentation = "https://github.com/kmaurinjones/AllMeans"
Repository = "https://github.com/kmaurinjones/AllMeans"
Issues = "https://github.com/kmaurinjones/AllMeans/issues"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --cov=AllMeans --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["AllMeans"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
